The norming continues
