+++ shell_exec/shell_exec.c
@@ -6,7 +6,7 @@
 /*   By: tmullan <tmullan@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/02/16 13:33:57 by tmullan       #+#    #+#                 */
-/*   Updated: 2021/05/20 15:47:35 by tmullan       ########   odam.nl         */
+/*   Updated: 2021/05/20 16:40:50 by tmullan       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -113,138 +113,6 @@ int	run_env(t_cmd *cmd, t_shell **ghost)
 		}
 		i++;
 	}
-	return (1);
-}
-
-int	export_replace(char *str, t_shell **ghost)
-{
-	int		i;
-	char	*var;
-
-	i = 0;
-	while (str[i])
-	{
-		if (str[i] == '=')
-			break ;
-		i++;
-	}
-	var = (char *)malloc(sizeof(char) * (i + 1)); //Protect the mallocs
-	ft_strlcpy(var, str, i + 1);
-	i = 0;
-	while ((*ghost)->env[i])
-	{
-		if (ft_strnstr((*ghost)->env[i], var, ft_strlen(var)))
-		{
-			if (ft_strchr(str, '='))
-			{
-				free((*ghost)->env[i]);
-				(*ghost)->env[i] = ft_strdup(str);
-			}
-			free(var);
-			return (0);
-		}
-		i++;
-	}
-	free(var);
-	return (1);
-}
-
-void	print_export(t_shell **ghost)
-{
-	int	i;
-	int	k;
-
-	i = 0;
-	k = 0;
-	while ((*ghost)->env[i])
-	{
-		ft_putstr_fd("declare -x ", 1);
-		if (ft_strchr((*ghost)->env[i], '='))
-		{
-			while ((*ghost)->env[i][k] != '"')
-			{
-				ft_putchar_fd((*ghost)->env[i][k], 1);
-				k++;
-				if ((*ghost)->env[i][k] == '=')
-				{
-					ft_putstr_fd("=\"", 1);
-					k++;
-					while ((*ghost)->env[i][k])
-					{
-						ft_putchar_fd((*ghost)->env[i][k], 1);
-						k++;
-					}
-					ft_putstr_fd("\"\n", 1);
-					k = 0;
-					i++;
-					break ;
-				}
-			}
-		}
-		else
-		{
-			ft_putstr_fd((*ghost)->env[i], 1);
-			ft_putchar_fd('\n', 1);
-			i++;
-		}
-	}
-}
-
-int	run_export(t_cmd *cmd, t_shell **ghost)
-{
-	int		i;
-	char	*str;
-
-	i = 0;
-	if (!cmd->args)
-	{
-		print_export(ghost);
-		return (1);
-	}
-	while (cmd->args->next)
-	{
-		str = cmd->args->content;
-		if (str[0] == '=')
-		{
-			cmd_notfound(cmd, EXPRT_FAIL, ghost, 0);
-			return (1);
-		}
-		while (str[i])
-		{
-			if (str[i] == '=')
-				break ;
-			if ((!ft_isalpha(str[i]) && (str[i] != '_' && str[i] != '$'
-			&& str[i] != '=' && str[i] != '/' && str[i] != '"' && str[i] != ' ')) || str[0] == '=')
-			{
-				cmd_notfound(cmd, EXPRT_FAIL, ghost, 0);
-				return (1);
-			}
-			i++;
-		}
-		if (export_replace(str, ghost))
-			(*ghost)->env = arr_addback((*ghost)->env, cmd->args->content);
-		cmd->args = cmd->args->next;
-	}
-	str = cmd->args->content;
-	if (str[0] == '=')
-	{
-		cmd_notfound(cmd, EXPRT_FAIL, ghost, 0);
-		return (1);
-	}
-	while (str[i])
-	{
-		if (str[i] == '=')
-			break ;
-		if ((!ft_isalpha(str[i]) && (str[i] != '_' && str[i] != '$'
-			&& str[i] != '=' && str[i] != '/' && str[i] != '"' && str[i] != ' ')) || str[0] == '=')
-		{
-			cmd_notfound(cmd, EXPRT_FAIL, ghost, 0);
-			return (1);
-		}
-		i++;
-	}
-	if (export_replace(str, ghost))
-		(*ghost)->env = arr_addback((*ghost)->env, cmd->args->content);
 	return (1);
 }
