+++ utils/struct_utils.c
@@ -6,7 +6,7 @@
 /*   By: ztan <ztan@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/03/08 11:25:13 by ztan          #+#    #+#                 */
-/*   Updated: 2021/05/20 15:14:28 by tmullan       ########   odam.nl         */
+/*   Updated: 2021/05/20 16:31:05 by tmullan       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
